refactor: Add user authentication to POST and DELETE routes in todos API

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserServerSession } from '@/auth/actions/auth-actions'
 import prisma from '@/lib/prisma'
 import { NextResponse } from 'next/server'
 import * as yup from 'yup'
@@ -29,12 +30,17 @@ const postSchema = yup.object({
 })
 
 export async function POST(req: Request) {
+  const user = await getUserServerSession()
+  if (!user) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+  }
+
   try {
     const { description, completed } = await postSchema.validate(await req.json())
 
     const todo = await prisma.todo.create({
       data: {
-        ...{ description, completed },
+        ...{ description, completed, userId: user.id },
       },
     })
 
@@ -45,10 +51,15 @@ export async function POST(req: Request) {
 }
 
 export async function DELETE(req: Request) {
+  const user = await getUserServerSession()
+  if (!user) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 })
+  }
   try {
     await prisma.todo.deleteMany({
       where: {
         completed: true,
+        userId: user.id,
       },
     })
     return NextResponse.json({ message: 'Deleted completed todos' })
